Address code review comments

+++ ambry-store/src/main/java/com.github.ambry.store/Log.java
@@ -83,7 +83,7 @@ public class Log implements Write, Read {
   }
 
   @Override
-  public void readInto(ByteBuffer buffer , int position) throws IOException {
+  public void readInto(ByteBuffer buffer , long position) throws IOException {
     fileChannel.read(buffer, position);
   }
 }
